Rename folder, Fix issues with form
